+++ doc/changelog.rst
@@ -3,6 +3,7 @@ Changelog
 
 Major changes to the pipeline since 2016-10-01 (see `Mommert 2017`_) are documented here.
 
+* 2017-03-16: implementation of `Pan-STARRS DR1`_ for the photometric calibration; currently, PP uses a home-built access of MAST at STScI, which is limited to catalog queries with a maximum cone radius of 0.5 deg; please note that this kind of query is rather slow compared to Vizier queries of SDSS or APASS
 
 * 2017-02-24: ``pillow`` is now a required python module; the pipeline now supports default distortion parameters for wide-field cameras
 
@@ -11,7 +12,7 @@ Major changes to the pipeline since 2016-10-01 (see `Mommert 2017`_) are documen
 
   
 .. _Mommert 2017: http://adsabs.harvard.edu/abs/2017A%26C....18...47M
-
+.. _Pan-STARRS DR1: http://panstarrs.stsci.edu/